refactored code structure following 3 layered architecture aswell as endpoint structure to be more inline with rest
