+++ data_access_layer/__init__.py
@@ -1,0 +1,1 @@
+from data_access_layer.firestore import *
