+++ presentation_layer/routers/__init__.py
@@ -1,0 +1,2 @@
+from presentation_layer.routers.router_image import *
+from presentation_layer.routers.router_path import *
